Moved formatIp function to the network client utils file
