+++ web/src/ConnectionsDataList.jsx
@@ -35,7 +35,8 @@ import {
   EOS_WIFI as WifiIcon
 } from "eos-icons-react";
 
-import { ConnectionTypes, formatIp } from "./client/network";
+import { ConnectionTypes } from "./client/network";
+import { formatIp } from "./client/network/utils";
 
 export default function ConnectionsDataList({ conns, onSelect }) {
   if (conns.length === 0) return null;
