+++ web/src/TargetIpsPopup.jsx
@@ -25,7 +25,7 @@ import Popup from "./Popup";
 
 import { useInstallerClient } from "./context/installer";
 import { useCancellablePromise } from "./utils";
-import { formatIp } from "./client/network";
+import { formatIp } from "./client/network/utils";
 
 export default function TargetIpsPopup() {
   const client = useInstallerClient();
